+++ frontend/controllers/PrestamosController.php
@@ -37,13 +37,20 @@ class PrestamosController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new PrestamosSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+        if(Yii::$app->user->can('modificarPrestamos'))
+        {
+
+            $searchModel = new PrestamosSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+    
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+
+        throw new NotFoundHttpException('Favor de Iniciar Sesión');
     }
 
     /**
@@ -54,9 +61,16 @@ class PrestamosController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+
+        if(Yii::$app->user->can('modificarPrestamos'))
+        {
+
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+
+        throw new NotFoundHttpException('Favor de Iniciar Sesión');
     }
 
     /**
@@ -66,51 +80,58 @@ class PrestamosController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Prestamos();
-        $modelsMateriales = [new Materiales];
 
-        if ($model->load(Yii::$app->request->post()) && $model->save())
+        if(Yii::$app->user->can('modificarPrestamos'))
         {
-            /*
-             * Se hace uso del  modelo "Model.php" , para el manejo
-             * múltiple de elementos
-            */
-            $modelsMateriales = Model::createMultiple(Materiales::classname());
-            Model::loadMultiple($modelsMateriales, Yii::$app->request->post());
 
-        
-            // validar los modelos
-            $valid = $model->validate();
-            $valid = Model::validateMultiple($modelsMateriales) && $valid;
+            $model = new Prestamos();
+            $modelsMateriales = [new Materiales];
+    
+            if ($model->load(Yii::$app->request->post()) && $model->save())
+            {
+                /*
+                 * Se hace uso del  modelo "Model.php" , para el manejo
+                 * múltiple de elementos
+                */
+                $modelsMateriales = Model::createMultiple(Materiales::classname());
+                Model::loadMultiple($modelsMateriales, Yii::$app->request->post());
+    
             
-            if ($valid) {
-                $transaction = \Yii::$app->db->beginTransaction();
-                try {
-                    if ($flag = $model->save(false)) {
-                        foreach ($modelsMateriales as $modelMateriales) {
-                            $modelMateriales->prest_id = $model->id;
-                            if (! ($flag = $modelMateriales->save(false))) {
-                                $transaction->rollBack();
-                                break;
+                // validar los modelos
+                $valid = $model->validate();
+                $valid = Model::validateMultiple($modelsMateriales) && $valid;
+                
+                if ($valid) {
+                    $transaction = \Yii::$app->db->beginTransaction();
+                    try {
+                        if ($flag = $model->save(false)) {
+                            foreach ($modelsMateriales as $modelMateriales) {
+                                $modelMateriales->prest_id = $model->id;
+                                if (! ($flag = $modelMateriales->save(false))) {
+                                    $transaction->rollBack();
+                                    break;
+                                }
                             }
                         }
+                        if ($flag) {
+                            $transaction->commit();
+                            return $this->redirect(['view', 'id' => $model->id]);
+                        }
+                    } catch (Exception $e) {
+                        $transaction->rollBack();
                     }
-                    if ($flag) {
-                        $transaction->commit();
-                        return $this->redirect(['view', 'id' => $model->id]);
-                    }
-                } catch (Exception $e) {
-                    $transaction->rollBack();
+                
                 }
-            
+               
             }
-           
+    
+            return $this->render('create', [
+                'model' => $model,
+                'modelsMateriales' => (empty($modelsMateriales)) ? [new Materiales] : $modelsMateriales
+            ]);
         }
 
-        return $this->render('create', [
-            'model' => $model,
-            'modelsMateriales' => (empty($modelsMateriales)) ? [new Materiales] : $modelsMateriales
-        ]);
+        throw new NotFoundHttpException('Favor de Iniciar Sesión');
     }
 
     /**
@@ -122,15 +143,22 @@ class PrestamosController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        if(Yii::$app->user->can('modificarPrestamos'))
+        {
+
+            $model = $this->findModel($id);
+    
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+    
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        throw new NotFoundHttpException('Favor de Iniciar Sesión');
     }
 
     /**
@@ -142,9 +170,16 @@ class PrestamosController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+        if(Yii::$app->user->can('modificarPrestamos'))
+        {
+
+            $this->findModel($id)->delete();
+    
+            return $this->redirect(['index']);
+        }
+
+        throw new NotFoundHttpException('Favor de Iniciar Sesión');
     }
 
     /**
@@ -160,6 +195,6 @@ class PrestamosController extends Controller
             return $model;
         }
 
-        throw new NotFoundHttpException('La página requerida no existe.');
+        throw new NotFoundHttpException('Favor de Iniciar Sesión');
     }
 }
